# app-service/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/products/{id}:
    get:
      tags:
      - product-rest-controller
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - product-rest-controller
      operationId: replaceProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - product-rest-controller
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/orders/{id}:
    get:
      tags:
      - order-rest-controller
      summary: Get an order by its id
      operationId: getOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    put:
      tags:
      - order-rest-controller
      summary: Modify the order with the given id
      operationId: replaceOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    delete:
      tags:
      - order-rest-controller
      summary: Delete the order with the given id
      operationId: deleteOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/v1/locations/{id}:
    get:
      tags:
      - location-rest-controller
      summary: Get the location with the given id
      operationId: getLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocationDTO"
    put:
      tags:
      - location-rest-controller
      summary: Modify the location with the given id
      operationId: replaceLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocationDTO"
    delete:
      tags:
      - location-rest-controller
      summary: Delete the location with the given id
      operationId: deleteLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocationDTO"
...
  /api/v1/orders/:
    get:
      tags:
      - order-rest-controller
      summary: Get a list of all orders
      operationId: getOrders
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDTO"
    post:
      tags:
      - order-rest-controller
      summary: Create a new order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/v1/locations/:
    get:
      tags:
      - location-rest-controller
      summary: Get all locations
      operationId: getLocations
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelLocationDTO"
    post:
      tags:
      - location-rest-controller
      summary: Create a new location
      operationId: createLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LocationDTO"
...
  /api/v1/orders/{id}/ticket:
    get:
      tags:
      - order-rest-controller
      summary: Get the invoice PDF of the order with the given id
      operationId: downloadTicket
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: byte
...
components:
  schemas:
    ImageBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
...
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/UserBasicDTO"
        image:
          $ref: "#/components/schemas/ImageBasicDTO"
        isSold:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          type: string
    UserBasicDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surnames:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
...
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        buyer:
          $ref: "#/components/schemas/UserBasicDTO"
        product:
          $ref: "#/components/schemas/ProductDTO"
        name:
          type: string
        surnames:
          type: string
        address:
          type: string
        apartment:
          type: string
        zipcode:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        phone:
          type: string
        creditCardNumber:
          type: string
          writeOnly: true
        creditCardExpiryDate:
          type: string
          writeOnly: true
        creditCardCVV:
          type: string
          writeOnly: true
        totalPrice:
          type: number
          format: double
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelLocationDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LocationDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
...